owner field added to hanqout, comment app completed

+++ comments/views.py
@@ -1,17 +1,19 @@
 from rest_framework.views import APIView
-from rest_framework.response import Response 
-from rest_framework import status 
+from rest_framework.response import Response
+from rest_framework import status
 from rest_framework.exceptions import NotFound
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.exceptions import PermissionDenied
 
 from .serializers.common import CommentSerializer
 from .models import Comment
 
 class CommentListView(APIView):
-    #permission_classes = (IsAuthenticated, )
+    permission_classes = (IsAuthenticated, )
     
     def post(self, request):
-        #request.data["owner"] = request.user.id 
-        comment_to_create = CommentSerializer(data=request.data) 
+        request.data["owner"] = request.user.id
+        comment_to_create = CommentSerializer(data=request.data)
         if comment_to_create.is_valid():
             comment_to_create.save()
             return Response(comment_to_create.data, status=status.HTTP_201_CREATED)
@@ -19,10 +21,12 @@ class CommentListView(APIView):
 
 class CommentDetailView(APIView):
 
-    def delete(self, _request, pk):
+    def delete(self, request, pk):
         try:
             comment_to_delete = Comment.objects.get(pk=pk)
         except Comment.DoesNotExist:
-            raise NotFound(detail="Comment not found")
+            raise NotFound()
+        if comment_to_delete.owner != request.user:
+            raise PermissionDenied()
         comment_to_delete.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
